add cppman

+++ tools/dep/ubuntu_init.sh
@@ -1,15 +1,23 @@
 #!/bin/bash
 
-[ -f  /etc/dpkg/dpkg.cfg.d/excludes ] && rm -rf /etc/dpkg/dpkg.cfg.d/excludes
-
 apt-get install -y gawk curl  zsh language-pack-zh-hans language-pack-zh-hans-base \
-    man manpages manpages-dev manpages-posix manpages-posix-dev \
     git unzip wget curl python-dev cscope  cmake gdb lsof \
     python3-pip python-pip shellcheck libtool-bin gettext sudo \
     automake m4 autoconf libtool build-essential cmake pkg-config lua5.2 wamerican \
-    autojump valgrind nmap iproute2 net-tools sshfs  clang-format
+    autojump  nmap iproute2 net-tools sshfs
 
 apt-get remove -y neovim exuberant-ctags
+
+if [ "Y$OPT_MAN" == "Yyes" ];then
+    [ -f  /etc/dpkg/dpkg.cfg.d/excludes ] && rm -rf /etc/dpkg/dpkg.cfg.d/excludes
+    apt-get install -y \
+    cppman man manpages manpages-dev manpages-posix manpages-posix-dev \
+fi
+
+if [ "Y$OPT_VIM_C" == "Yyes" ];then
+    apt-get install -y \
+    valgrind clang-format
+fi
 
 # install ripgrep
 ! which rg && \
